Rename the "Spec" column to "Path"
This <th> is used at all levels. At the top level most directories
correspond to a spec, but not all, in particular HTML is split across
directories. And except for in css/, tests and directories that aren't
at the top level never correspond to specs. So just call it path.