[processor] Implement WPTReport and tests (#157)
* [processor] Implement WPTReport and tests

WPTReport abstracts wptreport.json and provides some helper functions to
summarize and split the full report (the split output is needed by the
current frontend).

The code is largely based on
https://github.com/web-platform-tests/results-collection/blob/master/src/scripts/upload-wpt-results.py
but without any sharding support.

wptreport.py is also an executable (requires Python 3), a handy script
to summarize and/or split a wptreport JSON locally. See its help message
for detailed usage.

The plan is to run unit tests via `tox`, which takes care of creating
virtualenv and installing dependencies. The setup currently works
locally as long as Python 3 and tox are available. It hasn't been
integrated into Travis or the project Makefile yet; the integration will
happen soon later.