Silence gloud install script
The script prints out a few thousand lines of logs. They were put inside
a pair of Travis folding marks, but are still too noisy (e.g. folding
doesn't take effect when logs are still being streamed, or when Makefile
is used locally).

In the Dockerfile of results receiver, the stdout of this script is
directed to /dev/null, which seems to be a good compromise so far as
some of the more useful information is in stderr, so let's do this in
the project Makefile, too.