Always use '-' as the separator in report URLs
With https://github.com/web-platform-tests/data-migration/pull/6 , the
unshard script will also start to use '-' as the separator, which is
more consistent with the naming convention in the frontend serving
bucket (wptd) and design docs. And there is no need for the processor to
special case this any more.

This change has no user-facing effect. And existing historical reports
will be renamed in batch later.