Add TimeStart and TimeEnd to TestRun
First step of #310

Results processor will extract 'time_start' and 'time_end' from reports,
which are microseconds since epochs, convert them into ISO UTC time
strings, and send them to /api/run.

The API handler will unmarshal them into TimeStart and TimeEnd in
TestRun, which are two newly added fields in this PR. In the absence of
the two fields, the API handler will use the current timestamp as
TimeStart and TimeEnd.

CreatedAt continues to be the current timestamp when the API handler
receives the request, but the "retro" mode is now removed.