Make interop a sibling class of wpt-results
The two modules share many duplicate properties and methods, they can
and should be promoted to a common base class, test-runs. In addition to
improving code health, this also fixes some small UI glitches in the
interop view as the results view has been getting more attention and
many improvements only went there.

The refactor will also make it easier to fix #362.

Also skip the sha parameter in the interop view if it's the default
("latest") to keep the behaviour consistent with the results view.