[Refactor] Make interop a sibling class of wpt-results (#367)
The two modules share many duplicate properties and methods, they can and
should be promoted to a common base class, test-runs. In addition to improving
code health, the refactor will also make it easier to fix #362, and apply the
cleanup regarding query parameters in #355 to the interop view as well.

Also skip the sha parameter in the Go handler of the interop view if it's the
default ("latest") to keep the behaviour consistent with the results view.