Improve exceptions and docstrings in WPTReport
WPTReport now raises (subclasses of) WPTReportError if the input data is
invalid. A `finalize()` method is also added to compute summary and
generate various metadata; it guarantees that no WPTReportError will be
raised afterwards.

Now the user of WPTReport can easily wrap all report processing in a try
block to catch all data-related exceptions and returns a 2XX response to
remove the task from TaskQueue, because there is no point retrying these
deterministic errors caused by the report itself (unlike I/O errors,
which are not caught and will be retried).

Besides, various docstrings are improved.