Hide the diff column in single-file results
We don't have meaningful diff data to show in the single-file result
view (the API simply returns a pair of delta pass/total for the test
file without any subtest details). Besides, the diff column is currently
broken in the web component anyway (empty on prod and uncaught
exceptions on staging). Hence, we hide the diff column in the
single-file view but keep the parameter in URL so that when users go
back to the directory view the diff column will appear again.

In terms of implementation, this PR stops spoofing a diff run into the
`testRuns` array, which is awkward and forces all subclasses of
TestRunsBase to deal with the special case. Instead, the PR creates a
dedicated `diffRun` property that's only used in wpt-results.html (i.e.
the directory view).

Lastly, the PR also fixes a UI glitch where the passing/total numbers
might be wrapped into two lines by setting nowrap in the number cells.