File-level view data mirrors view structure
Fixes #494

Polymer was only updating table rows when their name column happened to change. This was a consequence of depending on `testRuns` and `subtestName`. In the case of the "Harness status" row, the `subtestName` didn't change and Polymer doesn't know that `testRuns[i].subtests.STATUS.*` changed, so so we observed #494.

This change replaces most of the `test-file-results*.html` properties with a `resultsTable` that contains an array of`{name, results: [{status, message}]` where the inner array is parallel to the `testRuns` array. This mirrors the structure of the views themselves and simplifies the binding logic, while also fixing #494.