Drop Selenium and use WebDrivers directly
In an attempt to improve the stability of our integration tests (#566),
this change drops the intermediary Selenium Standalone server, and uses
chromedriver/geckodriver directly in our tests to remove one layer of
indirection.

One observed downside of having Selenium Standalone in between is that
sometimes chromedriver/geckodriver processes, along with the browser
processes, are left behind. (c.f.
https://github.com/SeleniumHQ/selenium/issues/6434 ). Besides, now that
chromedriver/geckodriver are more spec compatible, we don't need
Selenium to smooth over the discrepancies.

Lastly, this also fixes #461 by the way.