Revert "Abstract basic auth in AppEngine APIs out of receiver package (#634)" (#638)
This reverts commit 2453ca7a16c358b2b6a790e2e15e8a73a579bc5c.

Reverts web-platform-tests/wpt.fyi#634

It turns out that the "shared logic" to be used by the receiver and upcoming spanner APIs cannot be shared because there are two Datastore SDKs: One for App Engine Standard and one for the rest of Google Cloud. Since the spanner API will be in App Engine Flexible Environment, it needs to use the latter SDK (whereas the existing receiver uses the App Engine Standard SDK).