Initial implementation of spanner API-as-service (#635)
This establishes the configuration and code for a handler that accepts authenticated requests to push runs in Datastore and GCS to Cloud Spanner. The request handler returns immediately after constructing a (composite) identifier for the request that can be used to track push logs. A `TODO` is included to implement Datastore/GCS pull and Cloud Spanner push.