[processor] Make the processor (almost) idempotent
We cannot make the results processor truely atomic & idempotent, because
even though RawResultsURL can be used as a UID, it is not the key and we
cannnot query by a property in a Datastore transaction.

This is the current best effort to avoid duplicate runs, which involves
two checks of existence of RawResultsURL: first before the recursive
upload (to avoid wasting time overwriting the same data), and second
before creating the run in Datastore.