Search cache: Add filter functions implementations of Plan (#859)
Plans on a sharded index are implemented by executing a slice of filters, or a ShardedFilter, in parallel. This change introduces filter functions that mirror query atoms. A sharded index creates a slice of filters, or ShardedFilter, that correspond to its shards during the binding phase, then the ShardedFilter can be executed during the execution phase.

This change also creates a placeholder file for forthcoming index+filter bind+execute tests.