Search cache: Add load-newer-runs update
This change implements a polling strategy to load newer runs (runs added after cache backfill started). It also integrates polling into the cache's production setup routine.

Since, for now, this is the only mechanism for receiving new runs, the default polling configuration is relatively aggressive: Load up to 10 new runs every 10 seconds. The frequency can be reduced if/when we implement best effort pushing of runs from the processor to the cache on the processor's critical path. The purpose of implemention such a scheme would be to ensure that runs are cached before they are committed to datastore and begin showing up in the wpt.fyi UI.