Search cache: Inject base query to ignore irrelevant tests
This change filters out tests that are missing from all test runs of interest when serving queries from the search cache. This is achieved by turning every query, q, in to AND(baseQuery, q), where baseQuery ensures that results filter out (sub)tests that do not appear in runs bound to the query from any of the four major browsers.