searchcache: When some runs missing, HTTP 422 + partial results
This change allows `searchcache` to return `HTTP 422` and partial results when at least one, but not all, runs needed to service a query are found in the cache.

This allows clients to render partial results and decide whether or not to retry the query later to get more complete results.