Ensure same runs: `q.PrepareUserQuery(runIDs)` and `idx.Bind(runs, q)` (#1041)
This change fault that occurs during query execution that can be traced to expanding a query with `PrepareUserQuery` using different run IDs than those that correspond to the runs that are later used to bind the query.

The change accumulates `ids` (for expansion) and `runs` (for binding) together, tracking only the runs that are resident in the `Index`. This ensures correct prep+bind+execute behaviour, including correct error handling if one of the runs is removed from the `Index` between `ids`/`runs` accumulation and one of prep|bind|execute.