Implement screenshot uploader (#1191)
* Upgrade Python dependencies

* [processor] Add wptscreenshot and tests

wptscreenshot is a multiprocessing batch uploader of screenshots (for
better performance), which warrants careful testing.

The test setup has a simple Flask server in the background that prints
out the number of screenshots it receives in each request. The tests are
more like end-to-end integration tests: it points WPTScreenshot at the
local test server to record all requests made by worker processes.

* Add a test for gzip

* Convert WPTScreenshot into a context manager

to ensure we always release all resources even in the case of
exceptions.