Fix the custom error type MultiError
The New method has to return an error interface instead of a concrete
type. This is not only in accordance with the best practice (errors are
checked by type assertions[1]), but also avoids a weird issue when
assigning the return value to a predeclared variable of type error:
(error)(nil) != (*MultiError)(nil).

Add a test to prevent regressing this subtle intentional behaviour.

Drive-by fix: remove unnecessary tags in browsers_test.go.

[1]: https://blog.golang.org/error-handling-and-go