Allow WCT & large tests to run on GitHub Actions
GitHub Actions runs Docker as root, which makes browsers unhappy. We add
a non-priviledged user "browser" (UID hardcoded to 9999) to the image
and add some checks to use "sudo -u browser" when necessary.

In addition, we need to disable Chrome's sandbox in both WCT and
WebDriver tests, because GitHub Actions does not start Docker in
priviledged mode.