Clean up Travis configurations
Now that we run all of our critical tests on GitHub Actions, which is
faster and more reliable than Travis, we can remove them from Travis to
make Travis finish faster (so that PRs don't get stuck in pending status
for that long).

We still run deployment and cleanup jobs on Travis. This commit also
cleans up the remaining job definitions a bit:
1. Upgrade from Ubuntu Xenial (16.04 LTS) to Bionic (18.04 LTS).
2. Change the deprecated keyword `matrix` to `jobs`.
3. Use conditional stages to simplify job definitions.
4. Run the cleanup job in a different stage to prevent race conditions.
5. Allow experimental Puppeteer tests to fail.