[UI] Add WebKitGTK as a product (#1601)
This adds WebKitGTK as a non-default product to the UI. It is selectable in add-browser
dialogs, and in the insights page, but is not default-selected for e.g. wpt.fyi/results

This change should not affect API endpoints. Note that this means that WebKitGTK will
not show up by default on wpt.fyi/runs for example; again, it can be added via the edit button.