Fix build: use Go modules properly in flex
This fixes the build for Go App Engine Flex services (announcer &
searchcache) by building the services outside of GOPATH in the builder
container to make sure go.mod is in effect.

Background:

Previously, we did not use go.mod so we needed to copy source code into
GOPATH inside the builder container for Flex services. We did not change
that setup when created go.mod.

However, the Go toolchain does NOT enable the module mode when running
inside GOPATH. The setup happened to continue working because we had
been using the latest major versions (same as the master branches) of
all dependencies. This setup finally broke when go-github released v29
and we haven't upgraded.

This change fixes the root cause (bonus: now we are sure all deps are
also pinned properly in announcer & searchcach). We will upgrade
go-github later.