[searchcache] Improve error handling
Previously we had a bunch of cases where the API handler did not return
after encountering fatal errors. In some cases, we would even try to
rewrite the status code of the response after writing out the error
message, which is impossible.

There is no bulletproof way to prevent missing returns, but it still
helps to extract the bulk of the handler into a new function that
explicitly returns an error which is then handled in one single place.