Add the missing Content-Type header back
This attempts to fix #1830.

Missing Content-Type: application/x-url-form-encoded when posting an
HTTP form is clearly a bug, but it is unclear why it only started to
surface after #1811.

A preventative measure is also added to return an error when the upload
payload is empty, along with some minor improvements to logging.