Fix version name errors when deploying master
We used the short SHA of the version name when deploying master.
However, the short SHA might happen to take the form of scientific
notation, e.g. 1e12345, in which case it'd be interpreted as a floating
point number by gcloud command and the deployment would fail.

Add a prefix "rev-" to the version name to prevent this from happening.