Improve logging for check_run and check_suite paths (#1955)
Previously we logged debug information for check_run immediately upon receiving them,
but check_suites only after filtering them for 'allowed' checks. This unifies the two codepaths
to both log immediately, so that we can better understand calls we receive.

In addition, this commit also fixes a misleading log line in handleCheckRunEvent (which used
'check_suite' instead of 'check_run'), and adds the GitHub App Name to the debug info.