Abstract Taskcluster tasks into their own struct
This is a prepatory refactor for the Taskcluster Checks migration.
Instead of using tcqueue.TaskDefinitionAndStatus directly, we extract
the information we actually need into a new taskInfo struct.

The goal is to ultimately share the logic in extractArtifactURLs and
processTaskclusterBuild between the GitHub Status and the (upcoming)
GitHub Checks paths.