Unbreak api/taskcluster/webhook_test.go build
The PR to introduce an EventInfo struct pulled the test code into its
own package. Just before it landed, another PR added a new test that was
not updated to handle being in a different package. Since we don't have
a commit queue both PRs landed cleanly, but the combination broke the
build for master.

This PR fixes the newly added test.