Use the input owner/repo when requesting check_runs for a check_suite
Previously we hardcoded the WPT owner and repository. This failed for
the wpt-tc-checks test repository, and would also provide a less than
useful error message if a non-WPT repository send us a check_suite
event. Instead, check specifically that we are receiving the event from
an allowed source, and then use it in the check_runs call.