[Taskcluster] Reject tasks that have non-success state
During the migration to the GitHub Checks API for Taskcluster, we
mistakenly interpreted the 'status' field of a CheckRun as equivalent to
the 'status' field of a Taskcluster task. This is not the case; in the
Taskcluster case 'status' captures the success/failure state of a task,
but for CheckRuns it merely captures the finished/not-finished state.

Instead, CheckRuns should use the 'conclusion' field to get the result
of the task.

To keep things compatible with the existing logic that assumes
Taskcluster status results, we map 'success' --> 'completed'. In the
future we should remove the old status code and switch to just using the
conclusion.

Issue: https://github.com/web-platform-tests/wpt.fyi/issues/2091