Avoid double-installing geckodriver
Previously the post-install script for wct would end up installing
chromedriver and geckodriver whenever you ran 'npm install'. For
_go_webdriver_test, we would then end up installing two versions of
geckodriver because the `geckodriver` rule in the Makefile also installs
a version of geckodriver. So that we control what versions of drivers
are installed, this commit defines NOSELENIUM in the Makefile to tell
wct not to do that.

This fixes the broken go_chrome_test, which had broken because the
double-geckodrivers caused a broken command-line execution (due to an
unexpected space in the middle of the commandline).

Fixes https://github.com/web-platform-tests/wpt.fyi/issues/2076