Fix Docker environment after #2252
Fixes #2259.

Make sure we start Cloud Datastore emulator in Docker as well, and plumb
the necessary flags (e.g. the port number) around.

Drive-by: fix a flaky WebDriver test. In api/interop.go, we use
`it.Next` to load the next entity into a struct that may not be zero,
in which case Datastore library will not zero it out first and may leave
stale data in the struct (e.g. if a slice field already has N items but
the entity to be loaded only has N-1 items in that field, the last item
will not be removed from the struct). The issue only surfaced after
PR #2252 as the IDs in Datastore emulator are no longer sequential.