[results-processor] Make conflicts easier to debug
Currently conflicts reported by the processor report the field that
conflicted but not the values that it saw, e.g.:

"Conflicting 'browser_version' found in the merged report"

This change adds the values, to better help debugging what conflict
occurred. Note that we bail on the first report that fails, so it is
possible that we could have (e.g.) 5 reports with 5 different
'browser_version's, and we would only report the first two after
this change.