[triage] Allow the triaging of a test, unrelated to a product
This adds support for triaging a test unrelated to a product (e.g.
Chrome). The goal is to allow the triaging of tests that are problematic
due to the test itself (e.g. if the test violates the spec or has bugs),
rather than problematic due to a browser bug.

To achieve this without changing the current data model, we use the
current support for an empty product. This is currently unused, but was
meant to be for URLs that affect all tests so matches well to the new
use-case. Metadata for this case are shown next to the test name rather
than in the results boxes, and similarly one triages test issues by
selecting the test name cell.