Remove obsolete 'experimentalAlignedExceptEdge' admin flag
This flag was used to fetch edge runs back when they were not tagged
with the experimental label. It is no longer useful, and is also
misleading (the name implies it might remove Edge from the default
query, which it does not do). As such, remove it.

This also makes util/populate_dev_data.go closer to what is shipped on
wpt.fyi