Remove obsolete 'experimentalAlignedExceptEdge' admin flag (#2433)
This flag was used to fetch edge runs back when they were not tagged
with the experimental label. It is no longer useful, and is also
misleading (the name implies it might remove Edge from the default
query, which it does not do).  See
https://github.com/web-platform-tests/wpt.fyi/pull/2431#issuecomment-790748838

This change also makes util/populate_dev_data.go closer to what is
shipped on wpt.fyi - the flags are updated and a new Edge experimental
test run is added for the local testing sha. The data for that run is a copy
of the Edge stable data for the same sha.

Finally, the `webdriver/` README.md file is extensively updated with more
description and a lot of debugging tips I've discovered over the last day!