Remove deprecated experimentalByDefault and experimentalAligned flags
Both of these flags have been default-on in wpt.fyi for many years, and
we are exceedingly unlikely to change them to be false at any point.
Removing them clarifies the intent of the code - and has also revealed
an existing inconsistency in our codebase (see Default() in
shared/test_run_filter.go).