[compat2021] Add a default summary-over-time view on the graph (#2488)
The UI change here is not massive, but requires substantially
refactoring how the page works. We now load the unified CSV rather than
individual CSV files for each feature. To aid with this and provide a
layer of caching, a specific data management class is introduced
(generalized from just chart data as it will eventually be used for the
summary scores at the top as well).

In addition to the above, this PR also improves the browser changelog
tooltip action. Loading changelogs automatically is now supported for
Firefox stable, and paper-dialogs pop-ups are added for Firefox Nightly
and Safari.