Remove Edge as a default product, but add breadcrumb link
After a lot of discussion, we have decided to remove Edge as a default
product from wpt.fyi due to overlap in browser engine between Chrome and
Edge. In a perfect world we would invest in per-user default products,
but there is no resourcing for that currently.

To try to minimize the disruption, we add a breadcrumb link back to
easily let users get Edge back on the page.

See https://github.com/web-platform-tests/wpt.fyi/issues/1519