Replace realpath with git rev-parse --show-toplevel
`realpath` was used in three places to get an absolute path to the repo
directory. We can use `git rev-parse --show-toplevel` for this instead
and avoid the dependency.

WPT_PATH was never used anywhere, and therefore removed.

In the pre-push script, commands.sh is sourced, so that will define
WPTD_PATH, so the redefinition can just be dropped.

Fixes https://github.com/web-platform-tests/wpt.fyi/issues/2743.