Allow the nicehttp library to take care of serving static files
Upon moving to the flexible environment, the handler section no longer
has an effect when deploying to app engine. That means the static files
can no longer be served. However, the nicehttp library has a built in
static file serve if GAE_DEPLOYMENT_ID env var is empty. This commit
sets that env var to empty string to enable that functionality when
deployed to gcloud.

Also, add a getAppEngineConfigPath function which reads
APP_ENGINE_CONFIG_PATH env var. if not provided use the default
previously used. This is needed so that in case the handlers are
different between staging and prod, it can use the right file