Remove Universal Analytics tracking (#2921)
* Remove Universal Analytics tracking

This commit should be merged in once we are ready to fully move away
from Universal Analytics. The Google Analytics 4 tracking codes are
already in place from the work on #2917. This commit just removes
Universal Analytics

Closes #2916

* Use staging GA4 instead of UA

With removing Universal Analytics, we can have the default one be the
staging environment