Ignore gcloud tests for dependabot & ignore certain npm packages (#3049)
For security reasons, dependabot builds do not have access to the same
secrets as a regular CI build. As a result, it cannot run gcloud tests
properly. This patch ignores gcloud tests for dependabot builds.

Also, configure dependabot to ignore certain packages for /webapp.
They were ignored in renovate previously. During the migration, they
were only ignored for webdriver/ but not webapp/. This fixes that.