Fix CI issues (#3088)
This PR helps one persistent CI issue and one intermittent CI issue.

The persistent issue is described in #3089. It is causing all golang related github actions to fail for PRs and on main.

The fix is to mark the directory as safe in the git config.
Also added a verbose flag to all go build commands to easily determine what is going on. Without it, the true error is not visible.
The intermittent issue is regarding go builds failing because the datastore emulator does not come up in 10 seconds

The fix/workaround is to extend the timeout from 10 seconds to 30 seconds when it is started during the medium tests