Attempt to fix the deploy to staging
The main conditional makes sure that the github.repository is always 'web-platform-tests/wpt.fyi'

Then we can check if the event comes from a PR event or non PR event.
- For the PR event, check if a fork and not dependabot
- For non PR event, we can let the event happen (which is only the push
event according to the events in the `on` section above). Even for the
push event, we can let dependabot be the actor since the PR has already
been verified by a human and it will have access to the credentials when
we merge it.

Inspiration:
- https://github.com/orgs/community/discussions/26409#discussioncomment-3251818

Fixes: #3237