Update the .python-version version to 3.9
Symptom: python dependabot updates were updating but not to the latest
available.

Cause: Dependabot currently looks at .python-version [1]. As a result,
as we keep using python 3.9+, we are at risk if we keep building for a
higher python while the packages are only compatible with 3.7.

Changes:
Only pinned the major.minor version because dependabot is always
updating the patch version [2].

[1] https://github.com/dependabot/dependabot-core/issues/1455
[2] https://github.com/web-platform-tests/wpt.fyi/blob/dc1c8c42900f1a7fc986294385cc2aacfab6535f/.github/dependabot.yml#L42-L43