Fix #3906: Implement a /api/checks/ endpoint for GitHub Actions
This is very much a derivative of the implementation used for Azure
Pipelines, perhaps unsurprising given the similarity between the two
systems, but is very much different enough to justify a totally
separate implementation.

Like Azure Pipelines, we rely on a notification from the workflow that
it is complete, rather than relying on webhooks from GitHub (as
Taskcluster does), as this allows the workflow to provide metadata to
endpoint, rather than having to hard code so many magic strings in
wpt.fyi (notably, the name of the artifacts we're looking for).