Upgrade to Debian Bookworm and Python 3.11 (#4028)
This commit upgrades the Debian distro to Bookworm, which requires an update in the Python (3.11) and Java (17) versions.

It also comes with an update to the Google Cloud CLI package names and repository secret installation (the recommended approach in the deprecation section of man apt-key). This is a prerequisite for upgrading to Go 1.23.

Had to run pip-compile requirements.in as part of the process.