Improve support for parsing WebKit Nightly versions and SHAs
webkitgtk and wpewebkit nightly versions have the following format:
`wpewebkit-2.49.1 (294826@main)` where `2.49.1 (294826@main)` is the
version.
This is the standard naming on the WebKit repository where each commit
on the main WebKit branch has a unique identifier of `number@main`.

However, the current version/sha parser expects the WPT Git SHA
identifier to be passed after an `@`.

This patch improves the parser to ignore the first `@main` in the
identifier string, allowing it to get the right version and also
the optional SHA.