Delete old versions after deployment (#4527)
* Delete old versions after deployment

This changes the order in which deletion of old versions of wpt.fyi is
deleted from the infrastructure. Prior to this change the deletion
happened at the beginning of the deployment. The deletion itself takes
many minutes per service, with prompts inbetween, so the time from start
of deplyment to the new image actually being available in the
infrastructure was longer than it needed to be. With this change, the
deployment starts right away, and once the deployment is done, we ask
whether old versions should be deleted and do those deletions w/o
prompting per service, meaning the person doing the deployment doesn't
need to sit and wait etc.

* Fix nesting of if statement.